Extract plot component
